Move event to preload utils and app.tsx
